changed count to int in PrimitiveIngredient and code cleanup

+++ src/main/java/at/hannibal2/skyhanni/features/misc/items/EstimatedItemValueCalculator.kt
@@ -891,7 +891,7 @@ object EstimatedItemValueCalculator {
                 val ingredient = PrimitiveIngredient(ingredients)
 
                 totalPrice += if (ingredient.isCoin()) {
-                    ingredient.count
+                    ingredient.count.toDouble()
                 } else {
                     ingredient.internalName.getPrice() * ingredient.count
                 }
